added drawString; working on EEPROM and other LCD commands
Sparkiduino 1.6.8.2 appears to be more efficient with memory and allows
more commands. This version tightens up the memory use of the library on
the Sparki, and implements drawLine(), readPixel(), and drawString. Also
on the Sparki side, I'm working on reading from and writing to arbitrary
EEPROM locations. On the python side, I've implemented sendIR() and
receiveIR() and drawString().