fixed motors issue; removed getBattery(); reduced memory; added EEPROM
This update fixes a longstanding problem with the motors() command --
though I'm not certain why. Somehow, variables were being switched in a
call to the motors() function. By splitting the call into an assignment
followed by a call, the function works.

The getBattery() function was removed as it was redundant with an
existing system call and didn't add any functionality

Various strings in the code were pruned or reduced in length to save RAM
on Sparki

The EEPROM read / write functions have been added, but they are not well
tested