squashed the arguments bug! updated the documentation
Finally figured out the root of the motors() bug -- the compiler does
not process the function calls in order. I am uncertain if it is simply
this version of the compiler, or a change to all versions. What was
happening was that if I had a series of function calls as the arguments
to a function call, they were not being called in the order written. For
example:
motors(getSerialInt(), getSerialInt(), getSerialFloat());
The ints are sent over the serial port in a particular order, but the
compiler was making the calls in the reverse order. I broke the calls
out into variable assignments, and now everything works.