added timer generator and additional timeout code
Added the timer function for compatibility with the Myro library. The
timer function is a generator which takes a duration as an argument and
yields the time taken since the first call. When the duration expires,
the iteration ends. It's primarily used to run a block of code for a
particular period of time, e.g.:
for x in timer(60):
if getObstacle() > 0:
backward(1,1)
will move the robot backward if it detects an obstacle in front of it
for 60 seconds

Also updated the release version of sparkiduino (to denote something
compile with 1.6.9.4)

Also added some additional code to handle timeouts on Macs. I do not
have a Mac to test with, so debugging in that area is limited to times
when I can access one.