added no_op and bluetooth address convenience functions
Macs continue to be a source of frustration, as the bluetooth connection
between the mac and the robot times out quickly. In an effort to prevent
this to some degree, I have added a "noop" to sparki. All it does is
receive a command over the bluetooth. It doesn't cause the robot to do
anything at all.

I have also added a set of internal functions to make it easier to store
a bluetooth address in the EEPROM. This is primarily for the convenience
of an instructor (i.e. me) who has lots of robots with different
addresses.

I do not have a mac, so this is presently untested on that platform.