added syncWait and get_client_start commands
Synchronization of multiple robots across multiple computers can be
challenging. The sync_lib attempts to make that a little easier by
providing a convenient library addition.

The syncWait() command allows the user to have the sparki wait while
giving noop commands so that the connection is maintained.

The get_client_start() command provides lower level access to the
library.