changed printDebug() to use python std lib logging module
Previously, the printDebug() command was the primary method of getting
debug messages out of the library. While that command worked well, in
the interest of making the sparki_learning library more pythonic, I'm
transitioning it to use the logging module.

Future updates in this 1.5 version line will be separating functions to
make the library more robot agnostic -- that is, I'm hoping to be able
to use the same basic API with different robots, so I will be separating
the library into multiple files to make that easier.