provided ping util to "unstick" ultrasoic sensor
The ultrasonic sensor on the Sparki seems to stop working for some robots. This may be due to an issue with the ultrasonic assembly used (see <http://therandomlab.blogspot.com/2015/05/repair-and-solve-faulty-hc-sr04.html>). There are hardware hacks which can fix the problem; however, there is also a potential software fix. The sparki_ping_reset_util.ino can be uploaded to the Sparki directly to attempt to address the problem. I have had some success in fixing a Sparki which was previously "blind".

Also, this version adds the ability for LCDdrawPixel to silently ignore out of bound pixels rather than attempting to correct them. This is done to permit drawings which may appear only partially on the screen.