added error handling for receipt of bad data
The initialization function (init) gets the name of the robot from the EEPROM on the Sparki -- but if the name has never been set, the EEPROM has garbage in the location where the name would be stored. This version changes getName() so that if there's an error getting the name of the Sparki, it will be ignored but will print an error message.

Thanks to github user chanakyan for reporting the bug and providing additional information surrounding it