added noop threading
Sparkis seem to suffer lots of disconnections between computer and robot. I believe that aggressive power saving is the culprit, because commands sent in relatively rapid succession do not cause issues.

This version adds a thread which continually sends noops. A semaphore is implemented to prevent more than one command being issued to the robot at a time. The semaphore is implemented at the logical command level instead of the serial level because some commands request information from the robot and then listen for the answer. The semaphore locks more code than is strictly speaking necessary in order to maintain logging consistency.

The version being uploaded right here is UNTESTED and should be considered alpha at best.