converted noops to a subclass of thread
In order to be able to run in interactive mode, there's the requirement that the noop thread remained running after termination of the main program. This version adapts a "StoppableThread" class found online to allow subclassing of a NoopThread. It has the same basic flow as the previous version, but the NoopThread itself may be stopped by an event. Whenever a new NoopThread is started by the library, it checks to see if there's an existing NoopThread, and if so terminates it through a stop event.